% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edescribe.R
\name{edescribe}
\alias{edescribe}
\title{edescribe}
\usage{

  edescribe(endpoint = "http://localhost:9090/termite/toolkit/edescribe.api",
  id)
}
\arguments{
\item{id}{A string representing a TERMite-compatible ID for a single named
entity.}
}
\description{
\code{autocomplete} passes user-specified information to the Ontology API
edescribe endpoint. Operation of this function requires a valid license to
TERMite provided by SciBite.
}
