% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/termite.R
\name{get_entity}
\alias{get_entity}
\title{Get entity information}
\usage{
get_entity(termite_home = "http://localhost:9090/termite", entity_id,
  entity_type)
}
\arguments{
\item{termite_home}{A string specifying the IP and API endpoint of TERMite.
Defaults to "http://localhost:9090/termite" representing a typical
installation.}

\item{entity_id}{A string representing an entity ID of interest.}

\item{entity_type}{A string representing an entity type of interest.}
}
\value{
TERMite JSON.
}
\description{
Entity lookup function, given an entity type (e.g. GENE, INDICATION) and entity ID (e.g. CSF1, D010024),
creates and runs GET call of the format: http://localhost:9090/termite/toolkit/tool.api?t=describe&id=INDICATION:D001249.
}
