% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/termite.R
\name{get_entity_hits_from_tsv}
\alias{get_entity_hits_from_tsv}
\title{Get entity hits from TERMite TSV}
\usage{
get_entity_hits_from_tsv(termite_tsv_response)
}
\arguments{
\item{termite_tsv_response}{TSV TERMite response object}
}
\value{
dataframe
}
\description{
Converts TERMite TSV response to dataframe of entity hits.
}
