% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/termite.R
\name{get_entity_hits_from_xml}
\alias{get_entity_hits_from_xml}
\title{Get entity hits from TERMite XML}
\usage{
get_entity_hits_from_xml(termite_xml_response)
}
\arguments{
\item{termite_xml_response}{XML TERMite response object}
}
\value{
dataframe
}
\description{
Converts TERMite XML response to dataframe of entity hits.
}
