% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/termite.R
\name{TERMite}
\alias{TERMite}
\title{TERMite}
\usage{
TERMite(endpoint = "http://localhost:9090/termite", input,
  inputFile = FALSE, inputFileFormat = NULL, outputFile = NULL,
  outputFormat = "json", VOCabs = NULL, options = NULL)
}
\arguments{
\item{endpoint}{A string specifying the IP and API endpoint of TERMite.
Defaults to "http://localhost:9090/termite" representing a typical
installation.}

\item{input}{A string representing either text to be processed with TERMite
or an absolute filepath to the file to be processed.}

\item{inputFile}{A boolean, with TRUE indicating that \code{input} should be
considered a filepath. Defaults to \code{FALSE}.}

\item{inputFileFormat}{A string representing the format of the file to be
processed. Requires that \code{file} be \code{TRUE}. Defaults to
\code{NULL}}

\item{outputFile}{A string representing the absolute filepath where output
will be written. Defaults to \code{NULL}, output will be written to
terminal.}

\item{outputFormat}{A string representing the desired format of output.
Defaults to "json".}

\item{VOCabs}{A character vector enumerating the VOCabs to be used for
annotation. Defaults to \code{NULL}, all primary VOCabs will be applied.}

\item{options}{A named list enumerating the runtime options to be used. Names
correspond to valid TERMite options and values correspond to valid option
settings.}
}
\description{
\code{TERMite} passes user-specified information to a TERMite API endpoint.
Operation of this function requires a valid license to TERMite provided by
SciBite.
}
