% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texpress.R
\name{texpress_hits_from_json}
\alias{texpress_hits_from_json}
\title{Get entity hits from TExpress JSON}
\usage{
texpress_hits_from_json(texpress_json_response)
}
\arguments{
\item{texpress_json_response}{JSON TERMite response object}
}
\value{
dataframe
}
\description{
Converts TExpress JSON response to dataframe of pattern hits.
}
